Step 0 results: the panel, the brackets, and two clusterers that were wrong
The full-sample panel: 5.9M daily rows, 8,177 cmc_ids, 2015-11 to 2025-06,
4,798 deaths, 1,048 tokens censored by falling below rank 2000 while still
alive, 1,976 still ranked at the end.

The rank-2000 pull depth never binds on B3's floor: 0 of 114 month-ends. That
was expected to bind in early years and does not, for two different reasons the
table separates. Before 2019 CoinMarketCap did not rank 2,000 tokens at all
(603 in 2016, 967 in 2017), so the binding constraint was CMC's coverage rather
than the chosen depth. From 2019 the ranking does reach 2,000 and the deepest
priced token sits far below the floor: $279 against a $1.9M floor in 2019,
$1.1M against $28.5M in 2025.

Perpetual coverage is the cap gradient the tradability verdict turns on. Share
of bracket members with a listed perp on Binance, Hyperliquid, dYdX or Deribit,
2016 to 2025: B0 runs 80-100% throughout, B1 climbs 35% to 87%, B2 climbs 11%
to 76%, B3 starts at 6% and reaches only 48%.

Median final-day volume of a dying token is $25.51 across all 4,798 deaths,
with 3,084 under $1,000, so the volume-floor delisting policy classifies 74.1%
as total losses against 100% under total_loss and 0% under the naive
last_price.

Two clusterers were then found to be measuring something other than what they
claimed, both caught by the same planted-block test.

BNCClustering solved eigh(A+ - A-, D_tot) for the smallest eigenvalues. Balance
Normalized Cut (Chiang, Whang, Dhillon, CIKM 2012) is eigh(D+ - A+ + A-,
D+ + D-). Dropping the D+ term only shifts the spectrum, so the old form
selected from the opposite end of it: on planted signed blocks that SPONGE and
the corrected form both recover exactly (adjusted Rand index 1.000), it scored
-0.02, worse than labelling at random.

SPONGEClustering.fit_symmetric whitened the same unnormalized pencil that fit
uses, which is a similarity transform of the same eigenproblem. It returned
identical labels and identical eigenvalues, so any "SPONGEsym" column was plain
SPONGE run a second time. It now builds the symmetrically normalized signed
Laplacians of Cucuringu et al. (AISTATS 2019), and the two spectra differ.

Committed derived tables total 50 MB. Return panels are float32: daily log
returns need nothing like float64 and the round-trip error is 2.4e-07, which
halved them from 56.6 MB to 33.3 MB. The 167 MB daily listing panel stays
gitignored and rebuildable.

+++ stat_arb/run_bracket_panel.py
@@ -336,7 +336,12 @@ def main(argv: list[str] | None = None) -> int:
             continue
         r = returns[cols]
         r.columns = [str(c) for c in r.columns]
-        r.to_parquet(out_dir / f"bracket_returns_{bracket}.parquet", compression="zstd")
+        # float32 halves the committed panel. Daily log returns live around
+        # 1e-3 to 1e0 and float32 carries ~7 significant digits, so the error
+        # accumulated over the whole 3,500-day sample is order 1e-5, far below
+        # anything that moves a Sharpe ratio.
+        r.astype("float32").to_parquet(
+            out_dir / f"bracket_returns_{bracket}.parquet", compression="zstd")
         stacked = pd.concat(
             {k: membership[bracket][k][cols] for k in SURVIVORSHIP_TREATMENTS}, axis=1)
         stacked.columns = [f"{a}|{int(b)}" for a, b in stacked.columns]
